app/requirements.txt
streamlit==1.24.0
mysql-connector-python==8.0.33
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT